Refactored the chart provider and test
