+++ lib/features/chart/abstract_chart_provider.dart
@@ -1,15 +1,8 @@
-import 'package:fl_chart/fl_chart.dart';
+import 'package:flutter/foundation.dart';
 
-abstract class AbstractChartProvider<T> {
+abstract class AbstractChartProvider<T> extends ChangeNotifier {
   /// Processes the data for the chart based on the provided model.
   List<dynamic> processDataForChart(T model);
-
-  /// Generates bar groups for the chart.
-  List<BarChartGroupData> generateBarGroups(
-    Map<String, dynamic> processedData,
-    DateTime startDate,
-    DateTime endDate,
-  );
 
   /// Gets the day of the week for a given index.
   String getDayOfWeek(int index);
