+++ lib/features/chart/base_chart_widget.dart
@@ -3,28 +3,79 @@ import 'package:flutter/material.dart';
 
 class BaseChartWidget extends StatelessWidget {
   final List<BarChartGroupData> barGroups;
-  final List<String> dayLabels;
+  final String Function(int) dayLabel;
+  final double yAxisInterval;
 
   const BaseChartWidget({
     super.key,
     required this.barGroups,
-    required this.dayLabels,
+    required this.dayLabel,
+    required this.yAxisInterval,
   });
 
   @override
   Widget build(BuildContext context) {
-    return BarChart(
-      BarChartData(
-        barGroups: barGroups,
-        titlesData: FlTitlesData(
-          leftTitles: const AxisTitles(
-            sideTitles: SideTitles(showTitles: true),
+    final maxY = barGroups
+            .map((group) => group.barRods.first.toY)
+            .fold(0.0, (max, y) => y > max ? y : max) +
+        5;
+    return AspectRatio(
+      aspectRatio: 1.5,
+      child: BarChart(
+        BarChartData(
+          alignment: BarChartAlignment.spaceAround,
+          maxY: maxY,
+          titlesData: FlTitlesData(
+            bottomTitles: AxisTitles(
+              sideTitles: SideTitles(
+                showTitles: true,
+                getTitlesWidget: (double value, TitleMeta meta) {
+                  return Padding(
+                    padding: const EdgeInsets.only(top: 8.0),
+                    child: Text(
+                      dayLabel(value.toInt()),
+                      style: const TextStyle(fontSize: 12),
+                    ),
+                  );
+                },
+              ),
+            ),
+            leftTitles: AxisTitles(
+              sideTitles: SideTitles(
+                showTitles: true,
+                getTitlesWidget: (double value, TitleMeta meta) {
+                  return Text(
+                    value % 1 == 0
+                        ? value.toInt().toString()
+                        : '', // Show only integer values
+                    textAlign: TextAlign.right,
+                    style: const TextStyle(fontSize: 12),
+                  );
+                },
+                reservedSize: 28, // Adjust for label width
+                interval: yAxisInterval, // Calculate interval dynamically
+              ),
+            ),
+            topTitles: const AxisTitles(),
+            rightTitles: const AxisTitles(),
           ),
-          bottomTitles: AxisTitles(
-            sideTitles: SideTitles(
-              showTitles: true,
-              getTitlesWidget: (value, meta) {
-                return Text(dayLabels[value.toInt()]);
+          borderData: FlBorderData(show: false),
+          barGroups: barGroups,
+          barTouchData: BarTouchData(
+            touchTooltipData: BarTouchTooltipData(
+              getTooltipItem: (
+                BarChartGroupData group,
+                int groupIndex,
+                BarChartRodData rod,
+                int rodIndex,
+              ) {
+                return BarTooltipItem(
+                  rod.toY.toInt().toString(),
+                  const TextStyle(
+                    color: Colors.white,
+                    fontWeight: FontWeight.bold,
+                  ),
+                );
               },
             ),
           ),
