+++ lib/features/counter/chart/chart.dart
@@ -1,3 +1,2 @@
 export 'chart_page.dart';
-export 'counter_chart_provider.dart';
 export 'heatmap_page.dart';
