+++ lib/features/counter/chart/chart_page.dart
@@ -2,54 +2,46 @@ import 'package:flutter/material.dart';
 
 import 'package:provider/provider.dart';
 
+import '../../chart/chart.dart';
 import '../counter.dart';
-import 'chart_widget.dart';
-import 'counter_chart_provider.dart';
 
-class ChartPage extends StatefulWidget {
+class ChartPage extends StatelessWidget {
   final CounterModel counter;
 
   const ChartPage({super.key, required this.counter});
 
   @override
-  State<ChartPage> createState() => _ChartPageState();
-}
-
-class _ChartPageState extends State<ChartPage> {
-  @override
-  void initState() {
-    super.initState();
-    loadData();
-  }
-
-  void loadData() {
-    Provider.of<CounterChartProvider>(
-      context,
-      listen: false,
-    ).processDataForChart(
-      Provider.of<CounterListProvider>(
-        context,
-        listen: false,
-      ).getCounter(widget.counter),
-    );
-  }
-
-  @override
   Widget build(BuildContext context) {
-    return const Padding(
-      padding: EdgeInsets.all(16.0),
+    return Padding(
+      padding: const EdgeInsets.all(16.0),
       child: Column(
         crossAxisAlignment: CrossAxisAlignment.stretch,
         children: <Widget>[
-          Text(
+          const Text(
             'Exercise Counts - Last 7 Days',
             style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
             textAlign: TextAlign.center,
           ),
-          SizedBox(height: 20),
-          ChartWidget(),
-          SizedBox(height: 20),
-          Text(
+          const SizedBox(height: 20),
+          Consumer<CounterChartProvider>(
+            builder: (context, chartProvider, child) {
+              final barGroups = chartProvider.processDataForChart(counter);
+              final maxY = barGroups
+                      .map((group) => group.barRods.first.toY)
+                      .fold(0.0, (max, y) => y > max ? y : max) +
+                  5;
+
+              return !chartProvider.isProcessed
+                  ? const CircularProgressIndicator()
+                  : BaseChartWidget(
+                      barGroups: barGroups,
+                      dayLabel: (value) => chartProvider.getDayOfWeek(value),
+                      yAxisInterval: calculateYAxisInterval(maxY),
+                    );
+            },
+          ),
+          const SizedBox(height: 20),
+          const Text(
             'Days are based on the last 7 days.',
             style: TextStyle(fontSize: 12, color: Colors.grey),
             textAlign: TextAlign.center,
@@ -57,5 +49,17 @@ class _ChartPageState extends State<ChartPage> {
         ],
       ),
     );
+  }
+
+  static double calculateYAxisInterval(double maxY) {
+    if (maxY <= 10) {
+      return 1;
+    } else if (maxY <= 50) {
+      return 5;
+    } else if (maxY <= 100) {
+      return 10;
+    } else {
+      return 20; // Or some other logic for larger maxY values
+    }
   }
 }
