+++ lib/features/timer/chart/chart.dart
@@ -1,4 +1,3 @@
 export 'chart_page.dart';
 export 'duration_interval.dart';
 export 'heatmap_page.dart';
-export 'timer_chart_provider.dart';
