+++ lib/main.dart
@@ -2,11 +2,10 @@ import 'package:flutter/material.dart';
 
 import 'package:provider/provider.dart';
 
-import 'features/counter/chart/chart.dart';
+import 'features/chart/chart.dart';
 import 'features/counter/counter.dart';
 import 'features/notification/notification_service.dart';
 import 'features/settings/settings.dart';
-import 'features/timer/chart/chart.dart';
 import 'features/timer/timer.dart';
 import 'routes.dart';
