+++ test/features/chart/counter_chart_provider_test.dart
@@ -1,22 +1,8 @@
 import 'package:clock/clock.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:growth_gauge/features/counter/chart/counter_chart_provider.dart';
+import 'package:growth_gauge/features/chart/chart.dart';
 import 'package:growth_gauge/features/counter/model/model.dart';
-
-final testCounter = CounterModel(
-  id: '1',
-  name: 'Test Counter',
-  count: 5,
-  description: 'Test Description',
-  logs: [
-    CounterLog(
-      id: 'log1',
-      action: 'Incremented',
-      timestamp: DateTime(2025, 4, 28),
-    ),
-  ],
-);
 
 void main() {
   late CounterChartProvider provider;
@@ -43,21 +29,7 @@ void main() {
       );
 
       // Act
-      provider.processDataForChart(counter);
-
-      // Assert
-      expect(provider.barGroups.isNotEmpty, true);
-    });
-
-    test('generateBarGroups', () {
-      // Arrange
-      final dailyCounts = {'2025-04-28': 5};
-      final startDate = DateTime(2025, 4, 27);
-      final endDate = DateTime(2025, 4, 30);
-
-      // Act
-      final barGroups =
-          provider.generateBarGroups(dailyCounts, startDate, endDate);
+      final barGroups = provider.processDataForChart(counter);
 
       // Assert
       expect(barGroups.isNotEmpty, true);
@@ -76,15 +48,5 @@ void main() {
       // Assert
       expect(day, 'Sun');
     });
-
-    test('processDataForChart processes data correctly', () {
-      // Act
-      provider.processDataForChart(testCounter);
-
-      // Assert
-      expect(provider.barGroups.isNotEmpty, true);
-    });
-
-    // Add more tests for other methods
   });
 }
