+++ test/features/chart/timer_chart_provider_test.dart
@@ -1,24 +1,9 @@
 import 'package:clock/clock.dart';
 import 'package:flutter_test/flutter_test.dart';
 
+import 'package:growth_gauge/features/chart/chart.dart';
 import 'package:growth_gauge/features/timer/chart/duration_interval.dart';
-import 'package:growth_gauge/features/timer/chart/timer_chart_provider.dart';
 import 'package:growth_gauge/features/timer/model/model.dart';
-
-final testTimer = TimerModel(
-  id: '1',
-  name: 'Test Timer',
-  interval: const Duration(minutes: 5),
-  description: 'Test Description',
-  logs: [
-    TimerLog(
-      id: 'log1',
-      action: 'Started',
-      timestamp: DateTime(2025, 4, 28),
-      interval: const Duration(minutes: 2),
-    ),
-  ],
-);
 
 void main() {
   late TimerChartProvider provider;
@@ -46,21 +31,8 @@ void main() {
       );
 
       // Act
-      provider.processDataForChart(timer, DurationInterval.minute);
-
-      // Assert
-      expect(provider.barGroups.isNotEmpty, true);
-    });
-
-    test('generateBarGroups', () {
-      // Arrange
-      final dailyDurations = {'2025-04-28': const Duration(minutes: 5)};
-      final startDate = DateTime(2025, 4, 27);
-      final endDate = DateTime(2025, 4, 30);
-
-      // Act
-      final barGroups = provider.generateBarGroups(
-          dailyDurations, startDate, endDate, DurationInterval.minute);
+      provider.interval = DurationInterval.minute;
+      final barGroups = provider.processDataForChart(timer);
 
       // Assert
       expect(barGroups.isNotEmpty, true);
@@ -81,23 +53,5 @@ void main() {
       // Assert
       expect(day, 'Sun');
     });
-
-    test('getChartTitle', () {
-      // Act
-      final title = provider.getChartTitle(DurationInterval.minute);
-
-      // Assert
-      expect(title, contains('Total Time Spent'));
-    });
-
-    test('processDataForChart processes data correctly', () {
-      // Act
-      provider.processDataForChart(testTimer, DurationInterval.minute);
-
-      // Assert
-      expect(provider.barGroups.isNotEmpty, true);
-    });
-
-    // Add more tests for other methods
   });
 }
